requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
httptools==0.1.2
pydantic==1.8.1
python-dotenv==0.17.1
PyYAML==5.4.1
starlette==0.13.6
typing-extensions==3.10.0.0
uvicorn[standard]
uvloop==0.15.2
watchgod==0.7
websockets==8.1
email-validator==1.1.2
SQLAlchemy==1.4.12
SQLAlchemy-Utils==0.37.0
async-exit-stack==1.0.1
async-generator==1.10
python-multipart
passlib[bcrypt]
pyjwt
